+++ CHANGELOG.rst
@@ -23,3 +23,4 @@ Latest
 * Fix bug with analysing relative imports from within __init__.py files.
 * Stop skipping analysing packages called ``migrations``.
 * Deal with invalid imports by warning instead of raising an exception.
+* Rename NetworkXBackedImportGraph -> ImportGraph.